New sqrt version

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: jonathan <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/16 17:03:53 by jonathan          #+#    #+#             */
-/*   Updated: 2020/12/24 12:29:26 by jonathan         ###   ########.fr       */
+/*   Updated: 2020/12/29 03:20:20 by jonathan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,7 +85,7 @@ void			ft_reverse_all_bytes(uint32_t *ptr);
 void			ft_reverse_byte(uint32_t *ptr, uint32_t idx);
 void			ft_reverse_each_byte(uint32_t *ptr);
 void			ft_set_bit(uint32_t *ptr, uint32_t idx);
-int				ft_sqrt(u_int32_t n);
+float			ft_sqrt(u_int32_t n);
 int				ft_strcasecmp(char const *s1, char const *s2);
 t_bool			ft_strcaseequ(char const *s1, char const *s2);
 char			*ft_strcat(char *dest, const char *src);
